Created controller implementation
